backend/requirements.txt
mysql-connector-python>=8.1.0
flask>=3.0.0
flask-cors>=4.0.0
stockfish>=3.28.0